+++ src/brightdata/datasets/crunchbase/__init__.py
@@ -1,0 +1,5 @@
+"""Crunchbase datasets."""
+
+from .companies import CrunchbaseCompanies
+
+__all__ = ["CrunchbaseCompanies"]