+++ src/brightdata/datasets/world_population/__init__.py
@@ -1,0 +1,5 @@
+"""World Population dataset."""
+
+from .countries import WorldPopulation
+
+__all__ = ["WorldPopulation"]